requirements.txt
asyncio
pandas
requests
websocket-client
openai
aiogram
pillow
numpy
openpyxl
xlsxwriter
python-dotenv 
openai-agents